requirements.txt
ultralytics
opencv-python
PyQt5
numpy
pillow
sqlalchemy
psycopg2-binary
python-docx
tqdm